+++ sapl/api/forms.py
@@ -3,19 +3,66 @@ import logging
 from django.db.models import Q
 from django.utils import timezone
 from django.utils.translation import ugettext_lazy as _
-from django_filters.filters import CharFilter, DateFilter
+from django_filters.filters import CharFilter, DateFilter, ModelChoiceFilter
 from django_filters.filterset import FilterSet
 from rest_framework import serializers
 
 from drfautoapi.drfautoapi import ApiFilterSetMixin
 from sapl.base.models import TipoAutor, Autor
 from sapl.parlamentares.models import Legislatura
+from sapl.utils import generic_relations_for_model
 
 logger = logging.getLogger(__name__)
 
 
 class SaplFilterSetMixin(ApiFilterSetMixin):
     pass
+
+
+class AutorFilterSet(SaplFilterSetMixin):
+    q = CharFilter(method='filter_q')
+    tipo = ModelChoiceFilter(queryset=TipoAutor.objects.all())
+
+    def filter_q(self, queryset, name, value):
+
+        query = value.split(' ')
+        if query:
+            q = Q()
+            for qtext in query:
+                if not qtext:
+                    continue
+                q_fs = Q(nome__icontains=qtext) | Q(
+                    tipo__descricao__icontains=qtext)
+
+                order_by = []
+
+                for gr in generic_relations_for_model(self._meta.model):
+                    sgr = gr[1]
+                    for item in sgr:
+                        if item.related_model != self._meta.model:
+                            continue
+                        flag_order_by = True
+                        for field in item.fields_search:
+                            if flag_order_by:
+                                flag_order_by = False
+                                order_by.append('%s__%s' % (
+                                    item.related_query_name(),
+                                    field[0])
+                                )
+                            # if len(field) == 3 and field[2](qtext) is not
+                            # None:
+                            q_fs = q_fs | Q(**{'%s__%s%s' % (
+                                item.related_query_name(),
+                                field[0],
+                                field[1]): qtext if len(field) == 2
+                                else field[2](qtext)})
+
+                q = q & q_fs
+
+            if q:
+                queryset = queryset.filter(q).order_by(*order_by)
+
+        return queryset.distinct()
 
 
 class AutoresPossiveisFilterSet(SaplFilterSetMixin):
